requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.1
img2pdf==0.6.0
Pillow==11.2.1
protobuf==6.31.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
Requests==2.32.3
